fix(eslint-remote-tester): unable to load TS configs with Jiti v2
This change updates the load logic for TS Configs, to use jiti 2, if we've detected the new api, and falling back to the old api, otherwise.

I also updated the version of jiti to v2.  I'd suggest also adding a ci step that installs jiti v1, and reruns the integration tests.